A fix.

+++ src/tiledata.c
@@ -210,7 +210,7 @@ uchar* tilemapflags;
                     else
                         Sprintf(eos(buf), ",1,1,0");
 
-                    Sprintf(eos(buf), ",%c,%d", def_monsyms[mons[i].mlet].sym, mons[i].mcolor);
+                    Sprintf(eos(buf), ",%c,%d", def_monsyms[(int)mons[i].mlet].sym, mons[i].mcolor);
 
                     //                    int tile_template_idx = (int)(mons[i].mflags6 & M6_MONSTER_TILE_TEMPLATE_MASK);
 //                    Sprintf(eos(buf), ",%d,%d,%s", (int)mons[i].mcolor, tile_template_idx, monster_tile_template_name_array[tile_template_idx]);
